fixed some warnings.
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@4999 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/DSUtil/DSMPropertyBag.cpp
@@ -440,7 +440,7 @@ STDMETHODIMP_(long) IDSMChapterBagImpl::ChapLookup(REFERENCE_TIME* prt, BSTR* pp
 
 	ChapSort();
 
-	int i = range_bsearch(m_chapters, *prt);
+	ptrdiff_t i = range_bsearch(m_chapters, *prt);
 	if (i < 0) {
 		return -1;
 	}
@@ -450,7 +450,7 @@ STDMETHODIMP_(long) IDSMChapterBagImpl::ChapLookup(REFERENCE_TIME* prt, BSTR* pp
 		*ppName = m_chapters[i].name.AllocSysString();
 	}
 
-	return i;
+	return (long)i;
 }
 
 STDMETHODIMP IDSMChapterBagImpl::ChapSort()
